+++ R/app_server.R
@@ -2,8 +2,6 @@
 #'
 #' @param input,output,session Internal parameters for {shiny}.
 #'     DO NOT REMOVE.
-#' @importFrom icecream ic_disable ic_enable
-#' @importFrom reactlog reactlog_module_server
 #' @noRd
 app_server <- function(input, output, session) {
   apply_server_settings()
@@ -62,5 +60,20 @@ app_server <- function(input, output, session) {
     params = dat_source[["params"]]
   )
 
-  if (getOption("shiny.reactlog", default = FALSE)) reactlog_module_server()
+  if (getOption("shiny.reactlog", default = FALSE) && is_installed("reactlog"))
+    reactlog::reactlog_module_server()
+}
+
+#' Apply some global settings on server startup
+#'
+#' @noRd
+apply_server_settings <- function() {
+  if (getOption("golem.app.prod") && is_installed("icecream")) {
+    icecream::ic_enable()
+    options(icecream.always.include.context = TRUE)
+  }
+
+  ggplot2::theme_set(hadex_ggtheme())
+
+  shinyhelper::observe_helpers(help_dir = app_sys("app/helpfiles"))
 }