+++ R/mod_data_load.R
@@ -44,7 +44,7 @@ mod_data_load_ui <- function(id){
     p("For the program to behave correctly, please make sure supplied file fulfills all requirements."),
     collapsible_card(
       title = "Show requirements",
-      htmlize_data(file_req), #TODO: maybe do it in some other way?
+      DT::datatable(file_req, options(dom = "t", pageLength = 15), width = "100%", height = "auto"),
       init_collapsed = TRUE
     ),
 
@@ -59,11 +59,10 @@ mod_data_load_ui <- function(id){
         mod_data_setup_ui(ns("setup"))
       ),
 
-      undisplay(
-        wellPanel(
-          id = "HaDeX-examiner-settings-panel",
-          mod_data_hdexaminer_ui(ns("hdexaminer"))
-        )
+      wellPanel(
+        id = "HaDeX-examiner-settings-panel",
+        mod_data_hdexaminer_ui(ns("hdexaminer")),
+        style = "display: none;"
       ),
       flex = c(NA, 1)
     )