+++ R/utils_constants.R
@@ -1,2 +1,5 @@
+#' Constants used in calculations
+#'
+#' @noRd
 MAX_TIME <- 99999
 PROTON_MASS <- 1.00727647