+++ NEWS.md
@@ -1,4 +1,10 @@
+# HaDeXGUI 0.1.1
+
+* Numerous documentation updates.
+* Reorganized code.
+* Created template for future plot tabs.
+
 # HaDeXGUI 0.1.0
 
 * Added a `NEWS.md` file to track changes to the package.
-* First version up-to-date with original HaDeX package
+* First version up-to-date with original HaDeX package.